Keep prefixed tags
Keep tag prefix "@" of the sqitch-plan output as it simplifies type
annotations and reading of the deploy scripts.  From now on the empty
string instead of None denotes untagged changes.  Only when generating
extension script filenames must the tag prefix be removed.  That's also
the place where the untagged HEAD should be denoted as version 'HEAD'.