Strip transaction control commands
Sqitch recommends BEGIN and COMMIT which are however not permitted in
extension scripts.  Filter out lines that contain commands BEGIN or
COMMIT.

The filter is not context-aware has at least one edge case where it
should not remove lines.  This edge case is covered by a unit test
marked with xfail.