Add @extschema@ placeholder
Using @extschema@ is already possible when justing building extension
scripts.  However, @extschema does not work with sqitch-deploy and,
depending on usage, results in syntax errors.

Use a custom placeholder that is a valid schema name to allow using
sqitch-deploy during development.  That placeholder is replaced with
@extschema@ when building the extension scripts.  The target database
of sqitch-deploy must define a schema matching the placeholder and
also put on the search_path.  Example with placeholder foobar:

    CREATE SCHEMA foobar;
    ALTER DATABASE test SET search_path = foobar;